Backend/HandbookScraper/requirements.txt
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fonttools==4.47.0
h11==0.14.0
idna==3.6
interchange==2021.0.4
kiwisolver==1.4.5
matplotlib==3.8.2
monotonic==1.6
networkx==3.2.1
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pansi==2020.7.3
Pillow==10.1.0
py2neo==2021.2.4
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
